feat(project): add task comments
Update the task page with a list of comments for that task, as well as a
form to add a comment to the task. Users can edit or delete their own
comments. Further, when a task is deleted, all associated comments are
deleted as well.

Finally, add a project member guard that verifies that the user is a
member of the current project for all project- and task-related
endpoints.